requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-filter==23.1
djangorestframework==3.14.0
geojson==2.5.0
idna==3.4
Markdown==3.4.3
pyowm==3.3.0
PySocks==1.7.1
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
